Add files via upload
I forgot to mention that I created the logo in photoshop and then added it the top div. This will be on every page along with the navbar and also the footer.

This page was a little tricky. I managed to get through it by using quite a bit of css to get what I wanted. It turned out nice. I added images and later on added links to each one.