Add files via upload
Copied over the usual html stuff.

Wanted to add a form to the site so I chose the contact us page as it made the most sense. This was a little tough as I had to learn how to code forms. I had to use em instead of px which made it a little confusing. 

There is no function behind the form but the look of the page is complete.