Add files via upload
This is my css. It has mostly classes inside that deal with divs. It has been open the entire time and is the one constant in making this website with Notepad++

I took the reset.css from marksheet.io. Wasn't sure if I was supposed to create one myself but by looking at that file I am thankful I didn't even try.